requirements.txt
flask
flask-socketio
opencv-python
numpy
torch
torchvision
gymnasium
scikit-learn
matplotlib
kagglehub
Pillow
gunicorn
eventlet
